bot/requirements.txt
discord.py==2.4.0
aiomysql==0.2.0
APScheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.0